requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
billiard==3.3.0.23
boto==2.48.0
boto3==1.4.4
botocore==1.5.80
celery==3.1.25
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
decorator==4.0.11
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11
django-celery==3.2.1
django-celery-results==1.0.1
django-common-helpers==0.9.1
django-crispy-forms==1.6.1
django-cron==0.5.0
django-filter==1.0.4
django-hitcount==1.2.2
django-model-utils==3.0.0
django-notifications-hq==1.3
django-storages==1.6.3
djangorestframework==3.5.4
docutils==0.13.1
easy-thumbnails==2.3
freeze==1.0.10
gunicorn==19.7.1
idna==2.6
imageio==2.1.2
jmespath==0.9.3
jsonfield==2.0.2
kombu==3.0.37
Markdown==2.6.8
nodeenv==1.1.2
numpy==1.13.0
oauthlib==2.0.2
olefile==0.44
packaging==16.8
Pillow==4.0.0
psycopg2==2.6.2
PyJWT==1.5.2
pyparsing==2.1.10
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.2
redis==2.10.5
requests==2.18.4
requests-oauthlib==0.8.0
rq==0.7.1
s3transfer==0.1.10
schedule==0.4.2
six==1.10.0
social-auth-core==1.4.0
tqdm==4.11.2
urllib3==1.22
vine==1.1.3
waitress==1.0.2
